chore: improve termination output
